pyproject.toml
[project]
name = "rag4ghart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipywidgets>=8.1.7",
    "modelscope>=1.28.0",
    "notebook>=7.4.4",
    "pandas-stubs>=2.3.0.250703",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pymilvus>=2.5.14",
    "tqdm>=4.67.1",
    "transformers[torch]==4.46.3",
]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true

[tool.uv.pip]
index-url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
